requirements.txt
streamlit
google-generativeai
deep-translator
pandas
python-dotenv
